Crypto-assignments/PublicKey_Crypto_Assignment_Kiruthika/requirements.txt
pycryptodome==3.19.0
cryptography==41.0.7
tinyec==0.4.0
python-gnupg==0.5.2